cache: Implement an on-disk pipeline library.
With VKD3D_SHADER_CACHE, we can add automatic serialization of pipeline
blobs to disk, even for games which do not make any use of GetCachedBlob
of ID3D12PipelineLibrary interfaces. Most applications expect drivers to
have some kind of internal caching.

For now, this is implemented as a fairly basic system where a disk
thread will manage a private ID3D12PipelineLibrary, and new PSOs are
automatically committed to this library. PSO creation will also consult
this internal pipeline library if applications do not provide their own
blob.

This design has some flaws in the read-write (first run) phase,
which is fairly IO hungry since we update the cache all in one go every time
(a design quirk we inherit from ID3D12PipelineLibrary here).
For this reason, we only update the cache periodically, either through
inactivity (game is done loading a scene most likely),
or periodically flushing the cache if games are in a spammy
phase. Ideally, we try to balance disk churn against responsiveness of
the cache.
Most games do not cleanly tear down their ID3D12Device, so we rarely get
to observe the clean teardown mechanism.

In the read-only scenario however, once caches are primed, we get the
optimal scenario where the cache can be mmaped directly from disk
without any copies or allocations.

Future iterations of this implementation may consider using a stream
format ala Fossilize which supports cheaper incremental updates to the
cache, but it needs a rework of ID3D12PipelineLibrary to handle both
styles of operation.

The cache is updated atomically through renames, which ensures that
even in hazardous scenarios, we should never end up with corrupt or
sliced caches.

Signed-off-by: Hans-Kristian Arntzen <post@arntzen-software.no>